--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/ScrimmageAuto.java
@@ -1,29 +1,0 @@
-package org.firstinspires.ftc.teamcode.Autonomous;
-
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-
-import org.firstinspires.ftc.teamcode.Robot;
-
-@Autonomous(name = "Scrimmage Auto", group = "1", preselectTeleOp = "MecanumTeleOp")
-public class ScrimmageAuto extends LinearOpMode {
-
-    private Robot wildWing = new Robot();
-
-    @Override
-    public void runOpMode() throws InterruptedException {
-        initRobot();
-        waitForStart();
-        if (opModeIsActive() && !isStopRequested()) {
-            wildWing.setPowers(0.5, 0.5, 0.5, 0.5, 0, 0);
-            sleep(1500);
-            wildWing.setPowers(-0.5, -0.5, -0.5, -0.5, 0, 0);
-            sleep(200);
-            wildWing.stopMoving();
-        }
-    }
-
-    public void initRobot() {
-        wildWing.init(hardwareMap, true);
-    }
-}
