+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Robot/Subsystem.java
@@ -1,0 +1,5 @@
+package org.firstinspires.ftc.teamcode.Robot;
+
+public abstract class Subsystem {
+    public abstract void runMotorsAndServos();
+}
